Displaying pr stats. Rest of files.
