+++ repo_health/index/static/client/src/components/base-component/base.template.html
@@ -12,7 +12,10 @@
          <md-card-title ng-if="$first">
           <md-card-title-text>
             <span class="md-headline">{{ $ctrl.stats[2 * i].raw_data }}</span>
-            <span class="md-subhead">{{ $ctrl.stats[2 * i + 1].raw_data }}</span>
+            <span class="md-subhead">
+                <b ng-if="$ctrl.stats[2 * i + 1].display_name">{{ $ctrl.stats[2 * i + 1].display_name }}</b>
+                {{ $ctrl.stats[2 * i + 1].raw_data }}
+            </span>
           </md-card-title-text>
         </md-card-title>
         <md-card-content layout="column" ng-if="!$first">
